Create script.js
The callback function takes three parameters, current element, index, and array(this). three arguments are required in different scenarios, sometimes array is needed or an index is needed but not all the time. Almost all the time current index is used but we have to declare when we are creating our own map function as we have to mimic the original map function.